FUCK OMNIWHEEL, old chassissubsystem works

+++ core/util/motor/DJIMotor.h
@@ -8,8 +8,8 @@
 #include <cmath>
 #include <string>
 
-constexpr double M3508_GEAR_RATIO = 3591.0 / 187.0;
-constexpr double M2006_GEAR_RATIO = 36.0;
+constexpr float M3508_GEAR_RATIO = 3591.0f / 187.0f;
+constexpr float M2006_GEAR_RATIO = 36.0f;
 
 constexpr int TICKS_REVOLUTION = 8192;
 constexpr int TIMEOUT_MS = 400;
